refactor(fs): `procfs` -> `proc`
