+++ kernel/src/file/fs/proc/proc_dir/mounts.rs
@@ -20,8 +20,8 @@
 
 use crate::{
 	file::{
-		fs::{procfs::get_proc_owner, Filesystem, NodeOps},
-		mountpoint, FileType, INode, Stat,
+        fs::{proc::get_proc_owner, Filesystem, NodeOps},
+        mountpoint, FileType, INode, Stat,
 	},
 	format_content,
 	process::pid::Pid,
